Change the code structure and update the message handling for Telegram.
